Not only is all != re not a valid shortcut, it also slows rak down a lot, and I cannot think of any case where it actually improves performance.
For piping in real time - rak just couldn't do that.

For huge files, rak would runs out of memory instead of just keeping a few lines at time.
Or if we only cared for the first N first - then just reading so much more data in
was pointless.

Not to mention regexps that had exponential backtracking - they are totally fine
if matched against one line at a time, but a huge file?

Anyway, now it's faster, corrected, and can handle streams.